requirements.txt
iris-pex-embedded-python
iris-embedded-python-wrapper
deepdiff
PyJWT
requests
irissqlcli